hybridRetrieval/hybridRetrieval/requirements.txt
requests
pinecone-text
langchain-openai
pinecone-client
nltk